requirements.txt
Flask
python-telegram-bot==13.15
schedule
python-dotenv
gunicorn
gspread
oauth2client
